symmetric aggregate handling
Generates the sql for sum/avg through a join, as well as fixing up count distinct in some cases.

Malloy has the concept of aggregate locality (ie users.avg()), but we're not supporting that just yet. For now, if you specify (avg(users.age)), we'll generate the equivalent of malloy's `users.avg(users.age)`