Support for long-running, multi-file analysis
My first version of multi-file was pretty hacky as it threw away all its work and re-read everything from disk each time. That was fine for the cli and evidence, but it's not great for a language server that wants to update diagnostics every few keystrokes.

This version is better, because it doesn't re-read or re-parse files that haven't changed. It does still re-analyze everything, because I don't wanna do dependency tracking yet.

It's also a bit smarter about scope. It knows that tables defined in one page shouldn't be visible in a different page, and that a subquery in queryA shouldn't be visible in queryB.

I also moved things a bit, `core.ts` is now the entrypoint and handles most of the high level api, while `analyze.ts` is just about analyzing.