Ditch web components
I was having a lot of difficulty getting these to work properly. I wanted light dom to make styling easier, but svelte doesn't like to mix that with <slot>. I tried svelte-retag, which is an alt implementation of web components, but it had issues with loading props into evidence, and some HMR weirdness. All in all, I think it would be possible but it just felt shaky.

This uses mdsvex to turn markdown into svelte, and then load that into the page. That makes it more heavily tied to svelte, and we have a compile step before we can show anything to the user.

The upside is makes components with children easier to build, provides better HMR, and lets us style things nicely.