Add param filling in `toSql`
When parsing we represent params as an Expression node, and look at the surrounding expressions to infer type.

When rendering sql, we take in a map of params, and will replace those param nodes with literal nodes. Because we want to reuse parses between queries, this will modify a structuredClone of the query.