sanitize md files
We want to ensure that you can't add random JavaScript to a Markdown file and have it execute. To do this we use sanitize-html to limit to just our graphene components, plus a few other safe ones from markdown.

I also ended up refactoring a bunch of the plugins involved in rendering markdown to its own file.