`alias: expr` syntax for computed columns
While a bit different from typical sql, this should make model much easier to scan, since the name of the thing you're defining is first rather than last.

We still support the old syntax, which makes the grammar somewhat convoluted. In the future I'd love to drop this old syntax.