Change serve back to fg by default
And drop it from graphene.md. Agents should never need to run the serve command directly, `check` will start it in the background automatically if needed.

This should make it feel a bit more normal for folks, plus if coding agents get good at managing background processes and tailing logs (which it seems like they would), we can leverage that.