Fix snowflake casing
My previous solution here didn't work for a couple of reasons:
1. The results object would come back from the database with uppercase column names, but G-SQL in the front end expects lowercase.
2. You couldn't actually refer to columns in your select things, so like order by wouldn't work properly. Or doing a where condition that referred to some select field wouldn't work.

This new solution feels a bit more elegant. We're basically using Molloy's aliasing feature, so every physical column is registered as its uppercase version or is changed to be its uppercase version and then aliased back as its lowercase version automatically. This means that all the G-SQL stuff works without a bunch of fancy traversal. We just have to make sure that we update all the physical columns even if they're nested within a join.